requirements.txt
django-admin-actions==0.1.1
Django==3.1.2
pandas==1.1.4
tushare==1.2.62
Scrapy==2.4.0
scrapy-djangoitem==1.1.1
